requirements.txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter and notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0

# Text processing
nltk>=3.8.0
spacy>=3.6.0
textstat>=0.7.0

# Model evaluation
evaluate>=0.4.0
rouge-score>=0.1.2

# Hyperparameter optimization
optuna>=3.2.0

# Data handling
huggingface-hub>=0.16.0
requests>=2.31.0

# Development tools
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0 